requirements.txt
asgiref==3.2.7
Django==3.0.5
django-rest-auth==0.9.5
djangorestframework==3.11.0
kafka-python==2.0.1
Pillow==7.1.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
